refactor: reorganize modules and rename files
