+++ Cargo.toml
@@ -8,10 +8,13 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0.66"
 clap = { version = "4.0.23", features = ["derive"] }
+clap-verbosity-flag = "2.0.0"
 home = "0.5.4"
+log = "0.4.17"
 reqwest = "0.11.12"
 serde = { version = "1.0.147", features = ["derive"] }
 serde_json = "1.0.87"
 serde_yaml = "0.9.14"
 tabled = { version = "0.10.0", features = ["derive"] }
 tracing = "0.1.37"
+tracing-subscriber = "0.3.16"
